requirements.txt
ollama
rich
matplotlib
flask
flask-socketio
pyyaml
requests
